Move permissions over to new app and expand code coverage
