+++ api/features/permissions.py
@@ -1,7 +1,6 @@
 from contextlib import suppress
 
 from django.shortcuts import get_object_or_404
-from rest_framework.generics import ListAPIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.viewsets import GenericViewSet
@@ -15,7 +14,7 @@ from environments.permissions.constants import (
     UPDATE_FEATURE_STATE,
     VIEW_ENVIRONMENT,
 )
-from features.models import Feature, FeatureExport, FeatureState
+from features.models import Feature, FeatureState
 from projects.models import Project
 from projects.permissions import CREATE_FEATURE, DELETE_FEATURE
 from projects.permissions import (
@@ -186,46 +185,3 @@ class CreateSegmentOverridePermissions(IsAuthenticated):
             permission=MANAGE_SEGMENT_OVERRIDES,
             environment=environment,
         )
-
-
-class FeatureImportPermissions(IsAuthenticated):
-    def has_permission(
-        self, request: Request, view: "WrappedAPIView"  # noqa: F821
-    ) -> bool:
-        if not super().has_permission(request, view):
-            return False
-
-        environment = Environment.objects.get(id=view.kwargs["environment_id"])
-        return request.user.is_environment_admin(environment)
-
-
-class CreateFeatureExportPermissions(IsAuthenticated):
-    def has_permission(
-        self, request: Request, view: "WrappedAPIView"  # noqa: F821
-    ) -> bool:
-        if not super().has_permission(request, view):
-            return False
-
-        environment = Environment.objects.get(id=request.data["environment_id"])
-        return request.user.is_environment_admin(environment)
-
-
-class DownloadFeatureExportPermissions(IsAuthenticated):
-    def has_permission(
-        self, request: Request, view: "WrappedAPIView"  # noqa: F821
-    ) -> bool:
-        if not super().has_permission(request, view):
-            return False
-
-        feature_export = FeatureExport.objects.get(id=view.kwargs["feature_export_id"])
-
-        return request.user.is_environment_admin(feature_export.environment)
-
-
-class FeatureExportListPermissions(IsAuthenticated):
-    def has_permission(self, request: Request, view: ListAPIView) -> bool:
-        if not super().has_permission(request, view):
-            return False
-
-        project = Project.objects.get(id=view.kwargs["project_id"])
-        return request.user.has_project_permission(VIEW_PROJECT, project)
